+++ src/Http/OctaneHandler.php
@@ -2,6 +2,7 @@
 
 namespace CacheWerk\BrefLaravelBridge\Http;
 
+use CacheWerk\BrefLaravelBridge\MaintenanceMode;
 use CacheWerk\BrefLaravelBridge\Octane\OctaneClient;
 
 use Illuminate\Http\Request;
@@ -22,7 +23,11 @@ class OctaneHandler extends HttpHandler
             SymfonyRequestBridge::convertRequest($event, $context)
         );
 
-        $response = OctaneClient::handle($request);
+        if (MaintenanceMode::active()) {
+            $response = MaintenanceMode::response($request)->prepare($request);
+        } else {
+            $response = OctaneClient::handle($request);
+        }
 
         return new HttpResponse(
             $response->getContent(),